Use external bitpack library
